Создание контроллера и представления для Сотрудников
