+++ WebStore/WebStore/Views/Employees/Index.cshtml
@@ -1,0 +1,10 @@
+﻿@model  IEnumerable<WebStore.Models.EmployeeView>
+@{
+    ViewBag.Title = "Сотрудники";
+}
+
+<h3>@ViewBag.Title</h3>
+
+<table>
+
+</table>
